Don't die if 'stty' isn't available
In order to be the most usable (without adding awkward checks
to the Makefile setup which users of the rest of the Log::Progress
dist might not want to bother with) just fall back to 80x25 dims
if 'stty' isn't usable.

Also, only install the SIGWINCH handler if stty succeeded.